refactor(Jackpot, Match, Ticket): 해당 클래스는 특정 로직을 담당하는 클래스이기 때문에 domains 폴더로 위치 변경
