+++ src/domains/Jackpot.js
